Reorganizacao diretorio hooks
